@@ -17,9 +17,10 @@ include=["src/py2of/logging.yaml"]
 Changelog = "https://github.com/ziky5/py2of/releases"
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "<3.13,>=3.9"
 click = ">=8.0.1"
 coloredlogs = "^15.0.1"
+numpy = "^1.26.1"
 
 [tool.poetry.dev-dependencies]
 Pygments = ">=2.10.0"
