@@ -1,11 +1,13 @@
 from dataclasses import dataclass
+from typing import Iterable, Sequence
 
 
-@dataclass()
+@dataclass
 class OfVector:
-    x: float
-    y: float
-    z: float
+    values: Sequence[int | float]
+
+    def __post_init__(self):
+        assert len(self.values) == 3
 
     def __str__(self) -> str:
-        return f"({self.x} {self.y} {self.z})"
+        return f"({self.values[0]} {self.values[1]} {self.values[2]})"
